requirements.txt
feedparser>=6.0.11,<7.0
python-dotenv>=1.0.1,<2.0
requests>=2.32.0,<3.0
trafilatura>=1.9.0,<2.0
openai>=1.99.0,<3.0
